Add close icon to episode details
Also made the FAB disappear as expected
